requirements.txt
Flask
Pillow
Werkzeug
python-magic
chardet
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
pillow-avif-plugin
opencv-python
numpy
pandas
tqdm
imagehash
requests